feat(qemu): rename package to `pytest-embedded-qemu`, with optional dependency `pytest-embedded-idf`
