+++ docs/plugins/modules.rst
@@ -6,7 +6,7 @@ Plugins API
 
    pytest_embedded
    pytest_embedded_serial
-   pytest_embedded_jtag
    pytest_embedded_serial_esp
    pytest_embedded_idf
-   pytest_embedded_qemu_idf
+   pytest_embedded_jtag
+   pytest_embedded_qemu
