--- docs/plugins/pytest_embedded_qemu_idf.rst
@@ -1,34 +1,0 @@
-pytest\_embedded\_qemu\_idf package
-===================================
-
-.. warning::
-
-   For now this plugin only support esp32 targets.
-
-Submodules
-----------
-
-pytest\_embedded\_qemu\_idf.dut module
---------------------------------------
-
-.. automodule:: pytest_embedded_qemu_idf.dut
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pytest\_embedded\_qemu\_idf.qemu module
----------------------------------------
-
-.. automodule:: pytest_embedded_qemu_idf.qemu
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pytest\_embedded\_qemu\_idf.plugin module
------------------------------------------
-
-.. automodule:: pytest_embedded_qemu_idf.plugin
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :exclude-members: pytest_addoption, pytest_plugin_registered
