--- pytest-embedded-qemu-idf/README.md
@@ -1,3 +1,0 @@
-# pytest-embedded-qemu-idf
-
-pytest embedded plugin for esp-idf project by qemu, not target chip
