+++ pytest-embedded-qemu/README.md
@@ -1,0 +1,3 @@
+# pytest-embedded-qemu
+
+pytest embedded plugin for qemu, not target chip
