+++ pytest-embedded-qemu/pytest_embedded_qemu/__init__.py
@@ -1,0 +1,1 @@
+DEFAULT_IMAGE_FN = 'flash_image.bin'
